SOLR-17284: Remove deprecated BlobRepository (#2447)
We have newer approaches such as the FileStore approach used by PackageManager.
